Refactoring: Got rid of metapackage and renamed ROS package to knowrob_moveit.
