+++ include/knowrob_moveit/planning_scene_server.hpp
@@ -13,8 +13,7 @@
  *   and/or other materials provided with the distribution.
  * 
  * * Neither the name of Institute for Artificial Intelligence nor the names of its
- *   contributors may be used to endorse or promote products derived from
- *   this software without specific prior written permission.
+ *   contributors may be used to endorse or promote products derived from *   this software without specific prior written permission.
  * 
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
@@ -28,11 +27,56 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
-#define KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP 
+#ifndef KNOWROB_MOVEIT_PLANNING_SCENE_SERVER_HPP 
+#define KNOWROB_MOVEIT_PLANNING_SCENE_SERVER_HPP 
 
-#include <knowrob_moveit_core/planning_scene.hpp>
-#include <knowrob_moveit_core/planning_scene_server.hpp>
-#include <knowrob_moveit_core/utils.hpp>
+#include <knowrob_moveit/planning_scene.hpp>
+#include <knowrob_moveit/utils.hpp>
+#include <knowrob_moveit/CheckCollisions.h>
 
-#endif // KNOWROB_MOVEIT_CORE_KNOWROB_MOVEIT_CORE_HPP
+namespace knowrob_moveit
+{
+  class PlanningSceneServer
+  {
+    public:
+  
+      PlanningSceneServer(const ros::NodeHandle& nh) : nh_( nh )
+      {}
+    
+      ~PlanningSceneServer() {}
+  
+      void start()
+      {
+        service_ = nh_.advertiseService("check_collisions", &PlanningSceneServer::callback, this);
+      }
+  
+    private:
+      ros::NodeHandle nh_;
+      ros::ServiceServer service_;
+      PlanningScene planning_scene_;
+
+      bool callback(knowrob_moveit::CheckCollisions::Request& request, 
+          knowrob_moveit::CheckCollisions::Response& response)
+      {
+        ROS_INFO("KnowRob-MoveIt check_collisions called.");
+      
+        try{
+          // TODO: use collision objects
+          response.contacts = planning_scene_.checkCollisions(request.urdf_model, request.srdf_model,
+              request.joint_states);
+        }
+        catch(const std::exception& e)
+        {
+          ROS_ERROR("%s", e.what());
+          return false;
+        }
+
+        ROS_INFO("KnowRob-MoveIt check_collisions finished.");
+ 
+        return true;
+      }
+  
+  };
+}
+
+#endif // KNOWROB_MOVEIT_PLANNING_SCENE_SERVER_HPP
