+++ include/knowrob_moveit/utils.hpp
@@ -28,26 +28,31 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <ros/ros.h>
-#include <knowrob_moveit_core/knowrob_moveit_core.hpp>
+#ifndef KNOWROB_MOVEIT_UTILS_HPP 
+#define KNOWROB_MOVEIT_UTILS_HPP 
 
-int main(int argc, char **argv)
+#include <moveit_msgs/ContactInformation.h>
+#include <moveit/collision_detection/collision_tools.h>
+
+// TODO: make argument const
+inline std::vector<moveit_msgs::ContactInformation> collisionResultToMsg(
+    collision_detection::CollisionResult& collision_result)
 {
-  ros::init(argc, argv, "planning_scene_server");
-  ros::NodeHandle nh("~");
+  std::vector<moveit_msgs::ContactInformation> result;
 
-  knowrob_moveit_core::PlanningSceneServer psserver(nh);
-  try
-  {
-    psserver.start();
-  }
-  catch(const std::exception& e)
-  {
-    ROS_ERROR("%s", e.what());
-    return 0;
-  }
+  for (collision_detection::CollisionResult::ContactMap::iterator it=collision_result.contacts.begin(); 
+       it!=collision_result.contacts.end(); ++it)
+    for (std::vector<collision_detection::Contact>::iterator it2=it->second.begin(); 
+         it2 != it->second.end(); ++it2)
+    {
+      moveit_msgs::ContactInformation msg;
+      collision_detection::contactToMsg(*it2, msg);
+      result.push_back(msg);
+    }
 
-  ros::spin();
+  ROS_INFO("Finished copying contacts.");
 
-  return 0;
+  return result;
 }
+
+#endif // KNOWROB_MOVEIT_UTILS_HPP
