--- knowrob_moveit/CMakeLists.txt
@@ -1,4 +1,0 @@
-cmake_minimum_required(VERSION 2.8.3)
-project(knowrob_moveit)
-find_package(catkin REQUIRED)
-catkin_metapackage()
