--- knowrob_moveit/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>knowrob_moveit</name>
-  <version>0.0.1</version>
-  <description>
-    A collection of packages that exposes moveit functionality to the knowrob system.
-  </description>
-
-  <maintainer email="georg.bartels@cs.uni-bremen.de">Georg Bartels</maintainer>
-
-  <license>BSD</license>
-
-  <author email="georg.bartels@cs.uni-bremen.de">Georg Bartels</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <run_depend>knowrob_moveit_core</run_depend>
-  <run_depend>knowrob_moveit_msgs</run_depend>
-
-  <export>
-    <metapackage/>
-  </export>
-</package>
